+++ packages/core/src/electron-preload/electron-windows-impl.ts
@@ -14,19 +14,23 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 // *****************************************************************************
 
-import { inject, injectable, postConstruct } from 'inversify';
-import { ElectronKeyboardLayout, ELECTRON_KEYBOARD_LAYOUT_IPC as ipc, NativeKeyboardLayout, TheiaIpcRenderer, IpcEvent, proxyable, proxy } from '../../electron-common';
+import { inject, injectable } from 'inversify';
+import { ElectronCurrentWindow, ElectronWindows, ELECTRON_WINDOWS_IPC as ipc, proxy, proxyable, TheiaIpcRenderer } from '../electron-common';
 
 @injectable() @proxyable()
-export class ElectronKeyboardLayoutImpl implements ElectronKeyboardLayout {
+export class ElectronWindowsImpl implements ElectronWindows {
 
-    @proxy() onKeyboardLayoutChanged: IpcEvent<NativeKeyboardLayout>;
+    @proxy() @inject(ElectronCurrentWindow)
+    currentWindow: ElectronCurrentWindow;
 
     @inject(TheiaIpcRenderer)
     protected ipcRenderer: TheiaIpcRenderer;
 
-    @postConstruct()
-    protected _postConstruct(): void {
-        this.onKeyboardLayoutChanged = this.ipcRenderer.createEvent(ipc.onKeyboardLayoutChanged);
+    @proxy() setMenuBarVisible(visible: boolean, windowName?: string): void {
+        this.ipcRenderer.send(ipc.setMenuBarVisible, visible, windowName);
+    }
+
+    @proxy() focusWindow(windowName: string): void {
+        this.ipcRenderer.send(ipc.focusWindow, windowName);
     }
 }
