+++ packages/filesystem/src/electron-preload/filesystem-preload-module.ts
@@ -14,21 +14,11 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 // *****************************************************************************
 
-import { proxy, proxyable, TheiaIpcRenderer } from '@theia/core/lib/electron-common';
-import { inject, injectable } from '@theia/core/shared/inversify';
-import { ElectronFileDialog, ELECTRON_FILE_DIALOG_IPC as ipc, OpenDialogOptions, SaveDialogOptions } from '../../electron-common';
+import { bindPreloadApi } from '@theia/core/lib/electron-common';
+import { ContainerModule } from '@theia/core/shared/inversify';
+import { ElectronFileDialog } from '../electron-common';
+import { ElectronFileDialogImpl } from './electron-file-dialog-impl';
 
-@injectable() @proxyable()
-export class ElectronFileDialogImpl implements ElectronFileDialog {
-
-    @inject(TheiaIpcRenderer)
-    protected ipcRenderer: TheiaIpcRenderer;
-
-    @proxy() showOpenDialog(cwd: string, options?: OpenDialogOptions): Promise<string[] | undefined> {
-        return this.ipcRenderer.invoke(ipc.showOpenDialog, cwd, options);
-    }
-
-    @proxy() showSaveDialog(cwd: string, options?: SaveDialogOptions): Promise<string | undefined> {
-        return this.ipcRenderer.invoke(ipc.showSaveDialog, cwd, options);
-    }
-}
+export default new ContainerModule(bind => {
+    bindPreloadApi(bind, ElectronFileDialog).to(ElectronFileDialogImpl).inSingletonScope();
+});
